Use ERB to template ConfigLoader files
Satisifies the failing test from the previous commit.

This has the advantage of not introducing another templating engine and the
ability to use Ruby to do anything in the template file. It also has the
disadvantage of being able to do *anything*, including loops and querying
external data sources. It will have to be left to the user's own discretion
whether they do crazy things that would render their configs unreproducible
between deployments.