Set NAT rule ID using an environment variable
Set the NAT rule ID using an environment variable which can be changed.

When writing the integration test for the NAT service, I found that
setting an abritrary value for the NAT rule `Id` triggered a HTTP 500
error on the vShield Edge Gateway:

> error: {:minorErrorCode=>"NETWORK_INTERNAL_SERVER", :message=>"Unable to
> perform this action. Contact your cloud administrator.",
> :majorErrorCode=>"500"}

When this occurred, I had used `Id` values of `500` and `501` for each
NAT rule.

I can't find any VMWare documentation[1] that explains the schema for
`Id`; as such I don't think this test would be very portable unless we
can define the rule IDs as needed.

Note that the rule IDs that we used (`65537` and `65538`) were taken
from a known working configuration.

[1]:
http://pubs.vmware.com/vcd-55/topic/com.vmware.vcloud.api.reference.doc_55/doc/types/NatRuleType.html