Use vCloud Tools Tester for edge gateway tests
Convert the edge gateway integration tests to use the [vCloud Tools
Tester](https://github.com/alphagov/vcloud-tools-tester) gem now that
it's available for use as of 9b02ef0e66e294f1f60ab4828da009bbd66664fb.

* * *

Also, bump the required version of vcloud-tools-tester to 0.0.6 so that
we can use the new test parameters set as of that version.